Refactored patch requests to be a part of their respective model-controllers
