+++ backend/src/controllers/userController.js
@@ -1,5 +1,5 @@
 const { request, response } = require("express");
-const { postUser, postLogin } = require("../models/userModel");
+const { postUser, postLogin, updateProviderStatus } = require("../models/userModel");
 
 exports.createUser = (request, response, next) => {
   const user = request.body;
@@ -19,3 +19,13 @@ exports.loginUser = (request, response, next) => {
     })
     .catch(next);
 };
+
+exports.patchProviderStatus = (request, response, next) => {
+  const { user_id } = request.params
+  const { isProvider } = request.body 
+  updateProviderStatus(Number(user_id), isProvider)
+  .then((user) => {
+      response.status(200).send(user)
+  })
+  .catch(next)
+}
